refactor: remove conftest.py, inline fixtures
Move temp_dir fixture into each test file to match
existing codebase pattern where each file is self-contained.
