+++ tests/cli/test_check_class_in_dir.py
@@ -1,6 +1,18 @@
 import os
+import shutil
+import tempfile
+from typing import Generator
+
+import pytest
 
 from cli import _check_class_in_dir
+
+
+@pytest.fixture
+def temp_dir() -> Generator[str, None, None]:
+    dir_path = tempfile.mkdtemp()
+    yield dir_path
+    shutil.rmtree(dir_path)
 
 
 class TestCheckClassInDir:
